# internal/api/v1/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Tailnet API",
    "title": "Tailnet API",
    "termsOfService": "https://github.com/sudosu404/providers/blob/main/LICENSE",
    "contact": {
      "name": "Yusing",
      "url": "https://github.com/sudosu404/providers/issues"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/sudosu404/providers/blob/main/LICENSE"
    },
    "version": "1.0"
  },
  "basePath": "/api/v1",
  "paths": {
...
    "/agent/list": {
      "get": {
        "description": "List agents",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "agent",
          "websocket"
        ],
        "summary": "List agents",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Agent"
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "x-id": "list",
        "operationId": "list"
      }
    },
...
    "/file/list": {
      "get": {
        "description": "List files",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "file"
        ],
        "summary": "List files",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ListFilesResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "x-id": "list",
        "operationId": "list"
      }
    },
    "/file/validate": {
      "post": {
        "description": "Validate file",
        "consumes": [
          "text/plain"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "file"
        ],
        "summary": "Validate file",
        "parameters": [
          {
            "enum": [
              "config",
              "provider",
              "middleware"
            ],
            "type": "string",
            "description": "Type",
            "name": "type",
            "in": "query",
            "required": true
          },
          {
            "description": "File content",
            "name": "file",
            "in": "body",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "File validated",
            "schema": {
              "$ref": "#/definitions/SuccessResponse"
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "417": {
            "description": "Validation failed",
            "schema": {}
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "x-id": "validate",
        "operationId": "validate"
      }
    },
...
    "/route/list": {
      "get": {
        "description": "List routes",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "route",
          "websocket"
        ],
        "summary": "List routes",
        "parameters": [
          {
            "type": "string",
            "description": "Provider",
            "name": "provider",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Route"
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "x-id": "routes",
        "operationId": "routes"
      }
    },
...
    "/route/providers": {
      "get": {
        "description": "List route providers",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "route",
          "websocket"
        ],
        "summary": "List route providers",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/RouteProvider"
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "x-id": "providers",
        "operationId": "providers"
      }
    },
...
  },
  "definitions": {
    "Agent": {
      "type": "object",
      "properties": {
        "addr": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "name": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "runtime": {
          "$ref": "#/definitions/agent.ContainerRuntime",
          "x-nullable": false,
          "x-omitempty": false
        },
        "version": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        }
      },
      "x-nullable": false,
      "x-omitempty": false
    },
...
    "Container": {
      "type": "object",
      "properties": {
        "agent": {
          "$ref": "#/definitions/Agent",
          "x-nullable": false,
          "x-omitempty": false
        },
        "aliases": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-nullable": false,
          "x-omitempty": false
        },
        "container_id": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "container_name": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "docker_host": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "errors": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "idlewatcher_config": {
          "$ref": "#/definitions/IdlewatcherConfig",
          "x-nullable": false,
          "x-omitempty": false
        },
        "image": {
          "$ref": "#/definitions/ContainerImage",
          "x-nullable": false,
          "x-omitempty": false
        },
        "is_excluded": {
          "type": "boolean",
          "x-nullable": false,
          "x-omitempty": false
        },
        "is_explicit": {
          "type": "boolean",
          "x-nullable": false,
          "x-omitempty": false
        },
        "is_host_network_mode": {
          "type": "boolean",
          "x-nullable": false,
          "x-omitempty": false
        },
        "labels": {
          "description": "for displaying in UI",
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "x-nullable": false,
          "x-omitempty": false
        },
        "mounts": {
          "description": "source:destination",
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "x-nullable": false,
          "x-omitempty": false
        },
        "network": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "private_hostname": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "private_ports": {
          "description": "privatePort:types.Port",
          "allOf": [
            {
              "$ref": "#/definitions/types.PortMapping"
            }
          ],
          "x-nullable": false,
          "x-omitempty": false
        },
        "public_hostname": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "public_ports": {
          "description": "non-zero publicPort:types.Port",
          "allOf": [
            {
              "$ref": "#/definitions/types.PortMapping"
            }
          ],
          "x-nullable": false,
          "x-omitempty": false
        },
        "running": {
          "type": "boolean",
          "x-nullable": false,
          "x-omitempty": false
        },
        "state": {
          "$ref": "#/definitions/container.ContainerState",
          "x-nullable": false,
          "x-omitempty": false
        }
      },
      "x-nullable": false,
      "x-omitempty": false
    },
    "ContainerImage": {
      "type": "object",
      "properties": {
        "author": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "name": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "sha256": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "tag": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "version": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        }
      },
      "x-nullable": false,
      "x-omitempty": false
    },
...
    "DockerConfig": {
      "type": "object",
      "required": [
        "container_id",
        "container_name",
        "docker_host"
      ],
      "properties": {
        "container_id": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "container_name": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "docker_host": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        }
      },
      "x-nullable": false,
      "x-omitempty": false
    },
    "ErrorResponse": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string",
          "x-nullable": true
        },
        "message": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        }
      },
      "x-nullable": false,
      "x-omitempty": false
    },
...
    "HealthCheckConfig": {
      "type": "object",
      "properties": {
        "disable": {
          "type": "boolean",
          "x-nullable": false,
          "x-omitempty": false
        },
        "interval": {
          "type": "integer",
          "x-nullable": false,
          "x-omitempty": false
        },
        "path": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "retries": {
          "description": "<0: immediate, >=0: threshold",
          "type": "integer",
          "x-nullable": false,
          "x-omitempty": false
        },
        "timeout": {
          "type": "integer",
          "x-nullable": false,
          "x-omitempty": false
        },
        "use_get": {
          "type": "boolean",
          "x-nullable": false,
          "x-omitempty": false
        }
      },
      "x-nullable": false,
      "x-omitempty": false
    },
    "HealthExtra": {
      "type": "object",
      "properties": {
        "config": {
          "$ref": "#/definitions/LoadBalancerConfig",
          "x-nullable": false,
          "x-omitempty": false
        },
        "pool": {
          "type": "object",
          "additionalProperties": {},
          "x-nullable": false,
          "x-omitempty": false
        }
      },
      "x-nullable": false,
      "x-omitempty": false
    },
...
    "HealthJSON": {
      "type": "object",
      "properties": {
        "config": {
          "$ref": "#/definitions/HealthCheckConfig",
          "x-nullable": false,
          "x-omitempty": false
        },
        "detail": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "extra": {
          "allOf": [
            {
              "$ref": "#/definitions/HealthExtra"
            }
          ],
          "x-nullable": true
        },
        "lastSeen": {
          "type": "integer",
          "x-nullable": false,
          "x-omitempty": false
        },
        "lastSeenStr": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "latency": {
          "type": "number",
          "x-nullable": false,
          "x-omitempty": false
        },
        "latencyStr": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "name": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "started": {
          "type": "integer",
          "x-nullable": false,
          "x-omitempty": false
        },
        "startedStr": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "status": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "uptime": {
          "type": "number",
          "x-nullable": false,
          "x-omitempty": false
        },
        "uptimeStr": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "url": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        }
      },
      "x-nullable": false,
      "x-omitempty": false
    },
...
    "HomepageItemConfig": {
      "type": "object",
      "properties": {
        "category": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "description": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "favorite": {
          "type": "boolean",
          "x-nullable": false,
          "x-omitempty": false
        },
        "icon": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "name": {
          "description": "display name",
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "show": {
          "type": "boolean",
          "x-nullable": false,
          "x-omitempty": false
        },
        "url": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "widget_config": {
          "allOf": [
            {
              "$ref": "#/definitions/widgets.Config"
            }
          ],
          "x-nullable": true
        }
      },
      "x-nullable": false,
      "x-omitempty": false
    },
...
    "IdlewatcherConfig": {
      "type": "object",
      "properties": {
        "depends_on": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-nullable": false,
          "x-omitempty": false
        },
        "docker": {
          "$ref": "#/definitions/DockerConfig",
          "x-nullable": false,
          "x-omitempty": false
        },
        "idle_timeout": {
          "description": "0: no idle watcher.\nPositive: idle watcher with idle timeout.\nNegative: idle watcher as a dependency.\tIdleTimeout time.Duration `json:\"idle_timeout\" json_ext:\"duration\"`",
          "allOf": [
            {
              "$ref": "#/definitions/time.Duration"
            }
          ],
          "x-nullable": false,
          "x-omitempty": false
        },
        "proxmox": {
          "$ref": "#/definitions/ProxmoxConfig",
          "x-nullable": false,
          "x-omitempty": false
        },
        "start_endpoint": {
          "description": "Optional path that must be hit to start container",
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "stop_method": {
          "$ref": "#/definitions/ContainerStopMethod",
          "x-nullable": false,
          "x-omitempty": false
        },
        "stop_signal": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "stop_timeout": {
          "$ref": "#/definitions/time.Duration",
          "x-nullable": false,
          "x-omitempty": false
        },
        "wake_timeout": {
          "$ref": "#/definitions/time.Duration",
          "x-nullable": false,
          "x-omitempty": false
        }
      },
      "x-nullable": false,
      "x-omitempty": false
    },
    "ListFilesResponse": {
      "type": "object",
      "properties": {
        "config": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-nullable": false,
          "x-omitempty": false
        },
        "middleware": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-nullable": false,
          "x-omitempty": false
        },
        "provider": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-nullable": false,
          "x-omitempty": false
        }
      },
      "x-nullable": false,
      "x-omitempty": false
    },
    "LoadBalancerConfig": {
      "type": "object",
      "properties": {
        "link": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "mode": {
          "$ref": "#/definitions/LoadBalancerMode",
          "x-nullable": false,
          "x-omitempty": false
        },
        "options": {
          "type": "object",
          "additionalProperties": {},
          "x-nullable": false,
          "x-omitempty": false
        },
        "weight": {
          "type": "integer",
          "x-nullable": false,
          "x-omitempty": false
        }
      },
      "x-nullable": false,
      "x-omitempty": false
    },
...
    "LogRetention": {
      "type": "object",
      "properties": {
        "days": {
          "type": "integer",
          "x-nullable": false,
          "x-omitempty": false
        },
        "keep_size": {
          "type": "integer",
          "x-nullable": false,
          "x-omitempty": false
        },
        "last": {
          "type": "integer",
          "x-nullable": false,
          "x-omitempty": false
        }
      },
      "x-nullable": false,
      "x-omitempty": false
    },
...
    "Port": {
      "type": "object",
      "properties": {
        "listening": {
          "type": "integer",
          "x-nullable": false,
          "x-omitempty": false
        },
        "proxy": {
          "type": "integer",
          "x-nullable": false,
          "x-omitempty": false
        }
      },
      "x-nullable": false,
      "x-omitempty": false
    },
...
    "RequestLoggerConfig": {
      "type": "object",
      "properties": {
        "buffer_size": {
          "description": "Deprecated: buffer size is adjusted dynamically",
          "type": "integer",
          "x-nullable": false,
          "x-omitempty": false
        },
        "fields": {
          "$ref": "#/definitions/accesslog.Fields",
          "x-nullable": false,
          "x-omitempty": false
        },
        "filters": {
          "$ref": "#/definitions/accesslog.Filters",
          "x-nullable": false,
          "x-omitempty": false
        },
        "format": {
          "type": "string",
          "enum": [
            "common",
            "combined",
            "json"
          ],
          "x-nullable": false,
          "x-omitempty": false
        },
        "path": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "retention": {
          "$ref": "#/definitions/LogRetention",
          "x-nullable": false,
          "x-omitempty": false
        },
        "rotate_interval": {
          "type": "integer",
          "x-nullable": false,
          "x-omitempty": false
        },
        "stdout": {
          "type": "boolean",
          "x-nullable": false,
          "x-omitempty": false
        }
      },
      "x-nullable": false,
      "x-omitempty": false
    },
    "Route": {
      "type": "object",
      "properties": {
        "access_log": {
          "allOf": [
            {
              "$ref": "#/definitions/RequestLoggerConfig"
            }
          ],
          "x-nullable": true
        },
        "agent": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "alias": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "container": {
          "description": "Docker only",
          "allOf": [
            {
              "$ref": "#/definitions/Container"
            }
          ],
          "x-nullable": true
        },
        "disable_compression": {
          "type": "boolean",
          "x-nullable": false,
          "x-omitempty": false
        },
        "excluded": {
          "type": "boolean",
          "x-nullable": true
        },
        "excluded_reason": {
          "type": "string",
          "x-nullable": true
        },
        "health": {
          "description": "for swagger",
          "allOf": [
            {
              "$ref": "#/definitions/HealthJSON"
            }
          ],
          "x-nullable": false,
          "x-omitempty": false
        },
        "healthcheck": {
          "$ref": "#/definitions/HealthCheckConfig",
          "x-nullable": false,
          "x-omitempty": false
        },
        "homepage": {
          "$ref": "#/definitions/HomepageItemConfig",
          "x-nullable": false,
          "x-omitempty": false
        },
        "host": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "idlewatcher": {
          "allOf": [
            {
              "$ref": "#/definitions/IdlewatcherConfig"
            }
          ],
          "x-nullable": true
        },
        "load_balance": {
          "allOf": [
            {
              "$ref": "#/definitions/LoadBalancerConfig"
            }
          ],
          "x-nullable": true
        },
        "lurl": {
          "description": "private fields",
          "type": "string",
          "x-nullable": true
        },
        "middlewares": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/types.LabelMap"
          },
          "x-nullable": true
        },
        "no_tls_verify": {
          "type": "boolean",
          "x-nullable": false,
          "x-omitempty": false
        },
        "path_patterns": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-nullable": true
        },
        "port": {
          "$ref": "#/definitions/Port",
          "x-nullable": false,
          "x-omitempty": false
        },
        "provider": {
          "description": "for backward compatibility",
          "type": "string",
          "x-nullable": true
        },
        "purl": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "response_header_timeout": {
          "type": "integer",
          "x-nullable": false,
          "x-omitempty": false
        },
        "root": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "rule_file": {
          "type": "string",
          "x-nullable": true
        },
        "rules": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/rules.Rule"
          },
          "x-nullable": false,
          "x-omitempty": false
        },
        "scheme": {
          "type": "string",
          "enum": [
            "http",
            "https",
            "tcp",
            "udp",
            "fileserver"
          ],
          "x-nullable": false,
          "x-omitempty": false
        },
        "ssl_certificate": {
          "description": "Path to client certificate",
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "ssl_certificate_key": {
          "description": "Path to client certificate key",
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "ssl_protocols": {
          "description": "Allowed TLS protocols",
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-nullable": false,
          "x-omitempty": false
        },
        "ssl_server_name": {
          "description": "SSL/TLS proxy options (nginx-like)",
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "ssl_trusted_certificate": {
          "description": "Path to trusted CA certificates",
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        }
      },
      "x-nullable": false,
      "x-omitempty": false
    },
    "RouteProvider": {
      "type": "object",
      "properties": {
        "full_name": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "short_name": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        }
      },
      "x-nullable": false,
      "x-omitempty": false
    },
...
    "SuccessResponse": {
      "type": "object",
      "properties": {
        "details": {
          "type": "object",
          "additionalProperties": {},
          "x-nullable": true
        },
        "message": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        }
      },
      "x-nullable": false,
      "x-omitempty": false
    },
...
    "accesslog.Fields": {
      "type": "object",
      "properties": {
        "cookies": {
          "$ref": "#/definitions/accesslog.FieldConfig",
          "x-nullable": false,
          "x-omitempty": false
        },
        "headers": {
          "$ref": "#/definitions/accesslog.FieldConfig",
          "x-nullable": false,
          "x-omitempty": false
        },
        "query": {
          "$ref": "#/definitions/accesslog.FieldConfig",
          "x-nullable": false,
          "x-omitempty": false
        }
      },
      "x-nullable": false,
      "x-omitempty": false
    },
    "accesslog.Filters": {
      "type": "object",
      "properties": {
        "cidr": {
          "$ref": "#/definitions/LogFilter-CIDR",
          "x-nullable": false,
          "x-omitempty": false
        },
        "headers": {
          "description": "header exists or header == value",
          "allOf": [
            {
              "$ref": "#/definitions/LogFilter-HTTPHeader"
            }
          ],
          "x-nullable": false,
          "x-omitempty": false
        },
        "host": {
          "$ref": "#/definitions/LogFilter-Host",
          "x-nullable": false,
          "x-omitempty": false
        },
        "method": {
          "$ref": "#/definitions/LogFilter-HTTPMethod",
          "x-nullable": false,
          "x-omitempty": false
        },
        "status_codes": {
          "$ref": "#/definitions/LogFilter-StatusCodeRange",
          "x-nullable": false,
          "x-omitempty": false
        }
      },
      "x-nullable": false,
      "x-omitempty": false
    },
    "agent.ContainerRuntime": {
      "type": "string",
      "enum": [
        "docker",
        "podman"
      ],
      "x-enum-varnames": [
        "ContainerRuntimeDocker",
        "ContainerRuntimePodman"
      ],
      "x-nullable": false,
      "x-omitempty": false
    },
...
    "container.ContainerState": {
      "type": "string",
      "enum": [
        "created",
        "running",
        "paused",
        "restarting",
        "removing",
        "exited",
        "dead"
      ],
      "x-enum-comments": {
        "StateCreated": "StateCreated indicates the container is created, but not (yet) started.",
        "StateDead": "StateDead indicates that the container failed to be deleted. Containers in this state are attempted to be cleaned up when the daemon restarts.",
        "StateExited": "StateExited indicates that the container exited.",
        "StatePaused": "StatePaused indicates that the container's current state is paused.",
        "StateRemoving": "StateRemoving indicates that the container is being removed.",
        "StateRestarting": "StateRestarting indicates that the container is currently restarting.",
        "StateRunning": "StateRunning indicates that the container is running."
      },
      "x-enum-descriptions": [
        "StateCreated indicates the container is created, but not (yet) started.",
        "StateRunning indicates that the container is running.",
        "StatePaused indicates that the container's current state is paused.",
        "StateRestarting indicates that the container is currently restarting.",
        "StateRemoving indicates that the container is being removed.",
        "StateExited indicates that the container exited.",
        "StateDead indicates that the container failed to be deleted. Containers in this state are attempted to be cleaned up when the daemon restarts."
      ],
      "x-enum-varnames": [
        "StateCreated",
        "StateRunning",
        "StatePaused",
        "StateRestarting",
        "StateRemoving",
        "StateExited",
        "StateDead"
      ],
      "x-nullable": false,
      "x-omitempty": false
    },
...
    "rules.Rule": {
      "type": "object",
      "properties": {
        "do": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "name": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "on": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        }
      },
      "x-nullable": false,
      "x-omitempty": false
    },
...
    "types.LabelMap": {
      "type": "object",
      "additionalProperties": {},
      "x-nullable": false,
      "x-omitempty": false
    },
    "types.PortMapping": {
      "type": "object",
      "additionalProperties": {
        "$ref": "#/definitions/container.Port"
      },
      "x-nullable": false,
      "x-omitempty": false
    },
    "widgets.Config": {
      "type": "object",
      "properties": {
        "config": {
          "x-nullable": false,
          "x-omitempty": false
        },
        "provider": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        }
      },
      "x-nullable": false,
      "x-omitempty": false
    }
  },
...
}